Rename ListCopyable to CopyableList
