+++ docs/App.jsx
@@ -36,7 +36,7 @@ import {
 	TextFieldRule,
 	TopBar,
 	DateField,
-	ListCopyable,
+	CopyableList,
 } from 'src'
 
 import IndexDoc from '../README.md'
@@ -65,7 +65,7 @@ import TextDoc from 'src/typography/Text/Readme.md'
 import TextFieldDoc from 'src/forms/TextField/Readme.md'
 import TextFieldRuleDoc from 'src/forms/TextFieldRule/Readme.md'
 import TopBarDoc from 'src/navigation/TopBar/Readme.md'
-import ListCopyableDoc from 'src/lists/ListCopyable/Readme.md'
+import CopyableListDoc from 'src/lists/CopyableList/Readme.md'
 import IconDoc from 'docs/icons/Readme.md'
 
 export default class App extends Component {
@@ -89,7 +89,7 @@ export default class App extends Component {
 		actionItem: '/lists/ActionItem',
 		contentItem: '/lists/ContentItem',
 		listItem: '/lists/ListItem',
-		listCopyable: '/lists/ListCopyable',
+		copyableList: '/lists/CopyableList',
 		spinner: '/loading/Spinner',
 		stepper: '/navigation/Stepper',
 		topbar: '/navigation/TopBar',
@@ -141,7 +141,7 @@ export default class App extends Component {
 				TextField,
 				TextFieldRule,
 				DateField,
-				ListCopyable,
+				CopyableList,
 				...Icons,
 			}
 
@@ -314,9 +314,9 @@ export default class App extends Component {
 								<TextDoc />
 							</Scroll>
 						</Screen>
-						<Screen className="doc-screen" displayMode="overlay" path={this.aliases.listCopyable}>
+						<Screen className="doc-screen" displayMode="overlay" path={this.aliases.copyableList}>
 							<Scroll>
-								<ListCopyableDoc />
+								<CopyableListDoc />
 							</Scroll>
 						</Screen>
 					</Router>
