+++ src/index.js
@@ -25,4 +25,4 @@ export { default as TextField } from './forms/TextField'
 export { default as TextFieldRule } from './forms/TextFieldRule'
 export { default as TopBar } from './navigation/TopBar'
 export { default as DateField } from './forms/DateField'
-export { default as ListCopyable } from './lists/ListCopyable'
+export { default as CopyableList } from './lists/CopyableList'
