+++ src/lists/CopyableList/index.test.js
@@ -1,7 +1,7 @@
 import { render } from '@testing-library/preact'
-import ListCopyable from '.'
+import CopyableList from '.'
 
-describe('UI/lists/ListCopyable', () => {
+describe('UI/lists/CopyableList', () => {
 
     const defaultProps = {
         title: 'Title',
@@ -27,20 +27,20 @@ describe('UI/lists/ListCopyable', () => {
     }
 
     it('Should show given title', () => {
-        const { getByTestId } = render(<ListCopyable {...defaultProps}/>);
+        const { getByTestId } = render(<CopyableList {...defaultProps}/>);
         const titleContentFound = getByTestId('heading').textContent === 'Title';
         expect(titleContentFound).toBeTruthy();
     })
 
     it('Should render given options', () => {
-        const { queryAllByTestId } = render(<ListCopyable {...defaultProps}/>);
+        const { queryAllByTestId } = render(<CopyableList {...defaultProps}/>);
         const allOptionsAreShown = queryAllByTestId('list-copyable-element').length === 3;
         expect(allOptionsAreShown).toBeTruthy();
     })
 
     it('Should display copy to clipboard button if hide property is false for given option', () => {
         const targetAttribute = 'list-copyable-copy-to-clipboard-button-button';
-        const { queryAllByTestId } = render(<ListCopyable {...defaultProps}/>);
+        const { queryAllByTestId } = render(<CopyableList {...defaultProps}/>);
         const option = queryAllByTestId('list-copyable-element')[0];
         const button = option.getElementsByTagName('button')[0];
         const copyToClipboardExist = button.getAttribute('data-testid') === targetAttribute
@@ -48,7 +48,7 @@ describe('UI/lists/ListCopyable', () => {
     });
     
     it('Should NOT display copy to clipboard button if hide property is true for given option', () => {
-        const { queryAllByTestId } = render(<ListCopyable {...defaultProps}/>);
+        const { queryAllByTestId } = render(<CopyableList {...defaultProps}/>);
         const option = queryAllByTestId('list-copyable-element')[1];
         const copyToClipboardNotFound = option.getElementsByTagName('button').length === 0;
         expect(copyToClipboardNotFound).toBeTruthy();
@@ -56,7 +56,7 @@ describe('UI/lists/ListCopyable', () => {
 
     it('Should display copy to clipboard button if hide property is NOT given in option', () => {
         const targetAttribute = 'list-copyable-copy-to-clipboard-button-button';
-        const { queryAllByTestId } = render(<ListCopyable {...defaultProps}/>);
+        const { queryAllByTestId } = render(<CopyableList {...defaultProps}/>);
         const option = queryAllByTestId('list-copyable-element')[2];
         const button = option.getElementsByTagName('button')[0];
         const copyToClipboardExist = button.getAttribute('data-testid') === targetAttribute
